requirements.txt
pip~=21.3.1
wheel~=0.37.1
setuptools~=60.2.0
python-dotenv~=0.20.0
googletrans==4.0.0rc1
python-slugify~=6.1.2
selenium~=4.4.0
youtube-search-python
youtube-search~=2.1.1
google-api-python-client~=2.56.0
pynput~=1.7.6
animu~=0.1.0
emoji~=1.4.0
wikipedia~=1.4.0
PyAutoGUI~=0.9.53
